feat(connection): multi-region support
Previously, when establishing a connection, a generic endpoint was
always chosen, and the user could not influence which region should be
preferred. This patch adds region support by allowing the user to pass
a perferred region to a connection. If a region cannot be found, the
process falls back to attempting to grab a generic region.